Fix termination, global consts, game order, SIGINT handle, close semaphores

+++ fedorov.dmitrii/lab2/connections/conn_fifo.cpp
@@ -15,18 +15,29 @@ std::string Conn::GetType ()
    return std::string("conn_fifo");
 }
 
+bool Conn::HasError ()
+{
+   return has_error;
+}
 
 Conn::Conn (int host_pid_, bool create) {
    owner = create;
    host_pid = host_pid_;
    std::string filename = std::string("/tmp/" + std::to_string(host_pid)).c_str();
    
-   unlink(filename.c_str());
    
-   int res = mkfifo(filename.c_str(), 0777);
-   if (res == -1) {
-      perror("mkfifo() ");
-      exit(EXIT_FAILURE);
+   if (owner) {
+      unlink(filename.c_str());
+      int res = mkfifo(filename.c_str(), 0777);
+      if (res == -1) {
+         perror("mkfifo() ");
+         RETURN_WITH_ERROR;
+      }
+   }
+
+   if ((desc = open(filename.c_str(), O_RDWR)) == -1) {
+      perror("open() ");
+      RETURN_WITH_ERROR;
    }
 
    if (owner) {
@@ -41,7 +52,7 @@ Conn::~Conn() {
    std::string filename = std::string("/tmp/" + std::to_string(host_pid)).c_str();
 
    if (owner) {
-      remove(filename.c_str());
+      unlink(filename.c_str());
       std::cout << "Connection: close fifo with name - " << filename << std::endl;
    }
 }
@@ -49,42 +60,28 @@ Conn::~Conn() {
 
 bool Conn::Read (void* buf, size_t count) {
    Msg msg;
-   int desc;   
-   std::string filename = std::string("/tmp/" + std::to_string(host_pid)).c_str();
 
-   if ((desc = open(filename.c_str(), O_RDONLY)) == -1) {
-      perror("open() ");
-      return false;
-   } else {
-      if (read(desc, &msg, count) == -1) {
-         perror("read() ");
-         close(desc);
-      } else {
-         *((Msg*)buf) = msg;
-         close(desc);
-         return true;
-      }
-   }
-   return false;
+   std::cout << "bef read"  << std::endl;
+ 
+   if (read(desc, &msg, count) == -1) {
+      perror("read() ");
+      RETURN_FALSE_WITH_ERROR;
+   } 
+
+   std::cout << "aft read"  << std::endl;
+ 
+     
+   *((Msg*)buf) = msg;
+   return true;   
 }
 
 
 bool Conn::Write (void* buf, size_t count) {
-   int desc;   
-   std::string filename = std::string("/tmp/" + std::to_string(host_pid)).c_str();
+   if (write(desc, buf, count) == -1) {
+      perror("write() ");
+      RETURN_FALSE_WITH_ERROR;
+   } 
 
-   if ((desc = open(filename.c_str(), O_WRONLY)) == -1) {
-      perror("open() ");
-      return false;
-   } else {
-      if (write(desc, buf, count) == -1) {
-         perror("write() ");
-         close(desc);
-      } else {
-         close(desc);
-         return true;
-      }
-   }
-   return false;
+   return true;
 }
